Fixing Visual Studio's pathetic merge attempt
