+++ Testing/TheQ.Utilities.CloudTools.Tests.Integration.Azure/ConcurrencyTests/GlobalMutexTests.cs
@@ -1,4 +1,4 @@
-﻿// <copyright file="GlobalLockTests.cs" company="nett">
+﻿// <copyright file="GlobalMutexTests.cs" company="nett">
 //      Copyright (c) 2015 All Right Reserved, http://q.nett.gr
 //      Please see the License.txt file for more information. All other rights reserved.
 // </copyright>
@@ -16,7 +16,7 @@ using System.Threading;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 using TheQ.Utilities.CloudTools.Azure;
-using TheQ.Utilities.CloudTools.Azure.GlobalLockFramework;
+using TheQ.Utilities.CloudTools.Azure.GlobalMutexFramework;
 
 namespace TheQ.Utilities.CloudTools.Tests.Integration.Azure.ConcurrencyTests
 {
@@ -31,12 +31,12 @@ namespace TheQ.Utilities.CloudTools.Tests.Integration.Azure.ConcurrencyTests
 			var result = string.Empty;
 			var consoleLog = new ConsoleLogService();
 			var lockProvider = new AzureLockStateProvider(consoleLog);
-			var factory = new AzureGlobalLockFactory((AzureBlobContainer) client.BlobClient.GetContainerReference("globallocktests1"), lockProvider, consoleLog);
+			var factory = new AzureGlobalMutexFactory((AzureBlobContainer) client.BlobClient.GetContainerReference("globalmutextests1"), lockProvider, consoleLog);
 
 			var thread1 = new Thread(
 				() =>
 				{
-					using (var globalLock = factory.CreateLock("test_lock"))
+					using (var globalMutex = factory.CreateLock("test_lock"))
 					{
 						result += "1";
 
@@ -47,11 +47,11 @@ namespace TheQ.Utilities.CloudTools.Tests.Integration.Azure.ConcurrencyTests
 				() =>
 				{
 					Thread.Sleep(3000); // Ensure it will enter later than the previous method
-					using (var globalLock = factory.CreateLock("test_lock")) result += "2";
+					using (var globalMutex = factory.CreateLock("test_lock")) result += "2";
 				});
 
 			// Act
-			client.BreakAnyLeases("globallocktests1", "test_lock");
+			client.BreakAnyLeases("globalmutextests1", "test_lock");
 			thread2.Start();
 			thread1.Start();
 			thread1.Join();
@@ -70,13 +70,13 @@ namespace TheQ.Utilities.CloudTools.Tests.Integration.Azure.ConcurrencyTests
 			var client = new CloudEnvironment();
 			var consoleLog = new ConsoleLogService();
 			var lockProvider = new AzureLockStateProvider(consoleLog);
-			var factory = new AzureGlobalLockFactory((AzureBlobContainer) client.BlobClient.GetContainerReference("globallocktests2"), lockProvider, consoleLog);
+			var factory = new AzureGlobalMutexFactory((AzureBlobContainer) client.BlobClient.GetContainerReference("globalmutextests2"), lockProvider, consoleLog);
 			var result = string.Empty;
 			var thread1 = new Thread(
 				() =>
 				{
 					bool isLocked;
-					using (var globalLock = factory.TryCreateLock("test_lock", out isLocked))
+					using (var globalMutex = factory.TryCreateLock("test_lock", out isLocked))
 					{
 						if (isLocked) result += "1";
 						Thread.Sleep(2000);
@@ -87,11 +87,11 @@ namespace TheQ.Utilities.CloudTools.Tests.Integration.Azure.ConcurrencyTests
 				{
 					Thread.Sleep(1000); // Ensure it will enter later than the previous method
 					bool isLocked;
-					using (var globalLock = factory.TryCreateLock("test_lock", out isLocked)) if (isLocked) result += "2";
+					using (var globalMutex = factory.TryCreateLock("test_lock", out isLocked)) if (isLocked) result += "2";
 				});
 
 			// Act
-			client.BreakAnyLeases("globallocktests2", "test_lock");
+			client.BreakAnyLeases("globalmutextests2", "test_lock");
 			thread2.Start();
 			thread1.Start();
 			thread1.Join();
@@ -111,14 +111,14 @@ namespace TheQ.Utilities.CloudTools.Tests.Integration.Azure.ConcurrencyTests
 			var result = string.Empty;
 			var consoleLog = new ConsoleLogService();
 			var lockProvider = new AzureLockStateProvider(consoleLog);
-			var factory = new AzureGlobalLockFactory((AzureBlobContainer) client.BlobClient.GetContainerReference("globallocktests3"), lockProvider, consoleLog);
+			var factory = new AzureGlobalMutexFactory((AzureBlobContainer) client.BlobClient.GetContainerReference("globalmutextests3"), lockProvider, consoleLog);
 
 
 			// Act
-			client.BreakAnyLeases("globallocktests3", "test_lock");
-			using (var globalLock = factory.CreateLock("test_lock")) result += "1";
+			client.BreakAnyLeases("globalmutextests3", "test_lock");
+			using (var globalMutex = factory.CreateLock("test_lock")) result += "1";
 
-			using (var globalLock = factory.CreateLock("test_lock")) result += "2";
+			using (var globalMutex = factory.CreateLock("test_lock")) result += "2";
 
 			// Assert
 			Assert.AreEqual("12", result);
@@ -136,16 +136,16 @@ namespace TheQ.Utilities.CloudTools.Tests.Integration.Azure.ConcurrencyTests
 			bool isLockedInner;
 			var consoleLog = new ConsoleLogService();
 			var lockProvider = new AzureLockStateProvider(consoleLog);
-			var factory = new AzureGlobalLockFactory((AzureBlobContainer) client.BlobClient.GetContainerReference("globallocktests4"), lockProvider, consoleLog);
+			var factory = new AzureGlobalMutexFactory((AzureBlobContainer) client.BlobClient.GetContainerReference("globalmutextests4"), lockProvider, consoleLog);
 
 			// Act
-			client.BreakAnyLeases("globallocktests4", "test_lock");
-			using (var globalLock = factory.TryCreateLock("test_lock", TimeSpan.FromSeconds(15), out isLocked))
+			client.BreakAnyLeases("globalmutextests4", "test_lock");
+			using (var globalMutex = factory.TryCreateLock("test_lock", TimeSpan.FromSeconds(15), out isLocked))
 			{
 				if (isLocked) result += "1";
 
 				Thread.Sleep(TimeSpan.FromSeconds(40));
-				using (var globalLockInner = factory.TryCreateLock("test_lock", out isLockedInner)) if (isLockedInner) result += "2";
+				using (var globalMutexInner = factory.TryCreateLock("test_lock", out isLockedInner)) if (isLockedInner) result += "2";
 			}
 
 			// Assert
@@ -164,15 +164,15 @@ namespace TheQ.Utilities.CloudTools.Tests.Integration.Azure.ConcurrencyTests
 			bool isLockedInner;
 			var consoleLog = new ConsoleLogService();
 			var lockProvider = new AzureLockStateProvider(consoleLog);
-			var factory = new AzureGlobalLockFactory((AzureBlobContainer) client.BlobClient.GetContainerReference("globallocktests5"), lockProvider, consoleLog);
+			var factory = new AzureGlobalMutexFactory((AzureBlobContainer) client.BlobClient.GetContainerReference("globalmutextests5"), lockProvider, consoleLog);
 
 			// Act
-			client.BreakAnyLeases("globallocktests5", "test_lock");
-			using (var globalLock = factory.TryCreateLock("test_lock", out isLocked))
+			client.BreakAnyLeases("globalmutextests5", "test_lock");
+			using (var globalMutex = factory.TryCreateLock("test_lock", out isLocked))
 			{
 				if (isLocked) result += "1";
 
-				using (var globalLockInner = factory.TryCreateLock("test_lock", out isLockedInner)) if (isLockedInner) result += "2";
+				using (var globalMutexInner = factory.TryCreateLock("test_lock", out isLockedInner)) if (isLockedInner) result += "2";
 			}
 
 			// Assert
@@ -190,16 +190,16 @@ namespace TheQ.Utilities.CloudTools.Tests.Integration.Azure.ConcurrencyTests
 			bool isLocked;
 			var consoleLog = new ConsoleLogService();
 			var lockProvider = new AzureLockStateProvider(consoleLog);
-			var factory = new AzureGlobalLockFactory((AzureBlobContainer) client.BlobClient.GetContainerReference("globallocktests6"), lockProvider, consoleLog);
+			var factory = new AzureGlobalMutexFactory((AzureBlobContainer) client.BlobClient.GetContainerReference("globalmutextests6"), lockProvider, consoleLog);
 
 			// Act
-			client.BreakAnyLeases("globallocktests6", "test_lock");
-			var globalLock = factory.CreateLock("test_lock");
+			client.BreakAnyLeases("globalmutextests6", "test_lock");
+			var globalMutex = factory.CreateLock("test_lock");
 			result += "1";
 
-			using (var globalLockInner = factory.TryCreateLock("test_lock", out isLocked)) if (isLocked) result += "2";
+			using (var globalMutexInner = factory.TryCreateLock("test_lock", out isLocked)) if (isLocked) result += "2";
 
-			globalLock.Unlock();
+			globalMutex.Unlock();
 
 			// Assert
 			Assert.AreEqual("1", result);
@@ -214,19 +214,19 @@ namespace TheQ.Utilities.CloudTools.Tests.Integration.Azure.ConcurrencyTests
 			var client = new CloudEnvironment();
 			var consoleLog = new ConsoleLogService();
 			var lockProvider = new AzureLockStateProvider(consoleLog);
-			var factory = new AzureGlobalLockFactory((AzureBlobContainer) client.BlobClient.GetContainerReference("globallocktests7"), lockProvider, consoleLog);
+			var factory = new AzureGlobalMutexFactory((AzureBlobContainer) client.BlobClient.GetContainerReference("globalmutextests7"), lockProvider, consoleLog);
 
 			// Act
-			using (var globalLock = factory.CreateLock("test_lock7", TimeSpan.FromSeconds(59)))
+			using (var globalMutex = factory.CreateLock("test_lock7", TimeSpan.FromSeconds(59)))
 			{
 			}
-			using (var globalLock = factory.CreateLock("test_lock7", TimeSpan.FromSeconds(60)))
+			using (var globalMutex = factory.CreateLock("test_lock7", TimeSpan.FromSeconds(60)))
 			{
 			}
 
 			try
 			{
-				using (var globalLock = factory.CreateLock("test_lock7", TimeSpan.FromSeconds(61)))
+				using (var globalMutex = factory.CreateLock("test_lock7", TimeSpan.FromSeconds(61)))
 				{
 				}
 
@@ -249,16 +249,16 @@ namespace TheQ.Utilities.CloudTools.Tests.Integration.Azure.ConcurrencyTests
 			var client = new CloudEnvironment();
 			var consoleLog = new ConsoleLogService();
 			var lockProvider = new AzureLockStateProvider(consoleLog);
-			var factory = new AzureGlobalLockFactory((AzureBlobContainer) client.BlobClient.GetContainerReference("globallocktests8"), lockProvider, consoleLog);
+			var factory = new AzureGlobalMutexFactory((AzureBlobContainer) client.BlobClient.GetContainerReference("globalmutextests8"), lockProvider, consoleLog);
 
 			// Act
-			client.BreakAnyLeases("globallocktests8", "test_lock");
-			var globalLock = factory.CreateLock("test_lock");
-			var lease = globalLock.CurrentLeaseId;
-			globalLock.LockAsync("test_lock").Wait();
+			client.BreakAnyLeases("globalmutextests8", "test_lock");
+			var globalMutex = factory.CreateLock("test_lock");
+			var lease = globalMutex.CurrentLeaseId;
+			globalMutex.LockAsync("test_lock").Wait();
 
 			// Assert
-			Assert.IsTrue(globalLock.CurrentLeaseId == lease);
+			Assert.IsTrue(globalMutex.CurrentLeaseId == lease);
 		}
 
 
@@ -270,17 +270,17 @@ namespace TheQ.Utilities.CloudTools.Tests.Integration.Azure.ConcurrencyTests
 			var client = new CloudEnvironment();
 			var consoleLog = new ConsoleLogService();
 			var lockProvider = new AzureLockStateProvider(consoleLog);
-			var factory = new AzureGlobalLockFactory((AzureBlobContainer) client.BlobClient.GetContainerReference("globallocktests9"), lockProvider, consoleLog);
+			var factory = new AzureGlobalMutexFactory((AzureBlobContainer) client.BlobClient.GetContainerReference("globalmutextests9"), lockProvider, consoleLog);
 
 			// Act
-			client.BreakAnyLeases("globallocktests9", "test_lock");
-			client.BreakAnyLeases("globallocktests9", "test_lock_alt");
-			var globalLock = factory.CreateLock("test_lock");
-			var lease = globalLock.CurrentLeaseId;
-			globalLock.LockAsync("test_lock_alt").Wait();
+			client.BreakAnyLeases("globalmutextests9", "test_lock");
+			client.BreakAnyLeases("globalmutextests9", "test_lock_alt");
+			var globalMutex = factory.CreateLock("test_lock");
+			var lease = globalMutex.CurrentLeaseId;
+			globalMutex.LockAsync("test_lock_alt").Wait();
 
 			// Assert II
-			Assert.IsTrue(globalLock.CurrentLeaseId != lease);
+			Assert.IsTrue(globalMutex.CurrentLeaseId != lease);
 		}
 
 
@@ -293,16 +293,16 @@ namespace TheQ.Utilities.CloudTools.Tests.Integration.Azure.ConcurrencyTests
 			var result = string.Empty;
 			var consoleLog = new ConsoleLogService();
 			var lockProvider = new AzureLockStateProvider(consoleLog);
-			var factory = new AzureGlobalLockFactory((AzureBlobContainer) client.BlobClient.GetContainerReference("globallocktests10"), lockProvider, consoleLog);
+			var factory = new AzureGlobalMutexFactory((AzureBlobContainer) client.BlobClient.GetContainerReference("globalmutextests10"), lockProvider, consoleLog);
 
 			var thread1 = new Thread(
 				() =>
 				{
-					using (var globalLock = factory.CreateLock("test_lock"))
+					using (var globalMutex = factory.CreateLock("test_lock"))
 					{
 						result += "1";
 						Thread.Sleep(5000);
-						client.BreakAnyLeases("globallocktests10", "test_lock");
+						client.BreakAnyLeases("globalmutextests10", "test_lock");
 					}
 
 					Thread.Sleep(3000);
@@ -312,11 +312,11 @@ namespace TheQ.Utilities.CloudTools.Tests.Integration.Azure.ConcurrencyTests
 				() =>
 				{
 					Thread.Sleep(3000); // Ensure it will enter later than the previous method
-					using (var globalLock = factory.CreateLock("test_lock")) result += "2";
+					using (var globalMutex = factory.CreateLock("test_lock")) result += "2";
 				});
 
 			// Act
-			client.BreakAnyLeases("globallocktests10", "test_lock");
+			client.BreakAnyLeases("globalmutextests10", "test_lock");
 			thread2.Start();
 			thread1.Start();
 			thread1.Join();
